requirements.txt
streamlit
python-docx
pypdf
google-genai